CAL-1320 : manage correctly checkbox visibility state of Task calendars
The visibility status of the calendars listed in the Project section in the Calendar application were not done. The status of the checkbox was always set as checked.
This fix manages this status correctly by checking the real state of the checkbox in order to add the right CSS class (as for others calendars types).
It also fixes the comparaison between the list of invisible calendars (stored in the settings as a comma-separated list of calendars ids) with the calendars ids. The previous comparaison was based on a basic string comparaison which led to bad results. The new comparaison is based on a list of compare exactly the ids.