CAL-1344 : avoid lazy initialization of the static field locale_ in remote calendar component (#119)
The field "locale_" of the webui component UIRemoteCalendar is static and updated depending on the locale of the current user. This can lead to conccurrency issues (2 users with different locale uses remote calendar conccurrently).
The fix uses a map to store the selectbox options of the component in different locales to avoid this potentiel issue.