CAL-1345 : avoid lazy initialisation of a static field in UISharedForm to avoid concurrency issues (#120)
The field INPUT_PERMISSION_OWNER_LABEL of the component UISharedForm is changed depending on the user locale  when the component is initialized. This can lead to concurrency issues when 2 users with different locales use this component concurrently.
This label is used as the placeholder of the input field for permissions.
This fix uses the native feature in HTML of input text placeholder (through the WebUI component UIFormStringInput) instead of simulating it ourself, and thus removes the lazy initialization of the field.