CAL-1445 : replace actions links in right-click context menu even for non-editable events (#140)
When right-clicking on event in the calendar, the pattern link of the context menu are replaced with the right links given the event properties (id, calId, ...).
This replacement was done only for editable event. Since task events are not editable, the links were not replaced.

This fix replaces also the links for non-editable events (tasks).