CAL-1438 : do not set seconds and milliseconds to 0 for end date of all day event (#141)
When an event is set as all day, its end date is stored with the value 23:59:00.000. This has changed compoared to the previous version which used the value 23:59:59.999.
This is due to the input event on the combobox which reset seconds and milliseconds to 0.
It leads to some regressions where we relied on this value.

This fix prevents the reset of seconds and milliseconds to 0 in case the event is an all day event.